Use fixed sizes in parser for stable stack use.
Also includes some breakdown of parse_operation so that we call
MolReader_RawTransaction_verify on a smaller stack and thus do not run
out of stack space.

+++ src/types.h
@@ -123,13 +123,13 @@ struct parsed_transaction {
     enum operation_tag tag;
     uint64_t total_fee;
     uint64_t amount; // 0 where inappropriate
-    uint32_t flags;  // Interpretation depends on operation type
+    uint8_t flags;  // Interpretation depends on operation type
     uint8_t source[20];
     uint8_t dao_source[20];
     uint8_t destination[20];
     uint8_t dao_destination[20];
     uint64_t dao_amount;
-    uint32_t group_input_count;
+    uint8_t group_input_count;
 };
 
 // Maximum number of APDU instructions
